add arduino component
This component is optional, this can be accomplished with any embedded system that has extra GPIO. In the case you don't have extra IO to control a couple of LEDs and the RTC you can use this. This is also useful if you cannot easily recompile your kernel. Since this requires no extra support except serial.